requirements.txt
MLflow==2.22.1
Flask==3.1.2
Flask_Cors==6.0.1
dvc==3.63.0
zenml==0.84.3
kagglehub==0.3.13
boto3==1.40.26
PyYAML==6.0.2
requests==2.32.5
numpy==2.3.2
pandas==2.3.2
scikit-learn==1.7.1
scipy==1.16.1
dagshub==0.6.3
xgboost==3.0.5
catboost==1.2.8
lightgbm==4.6.0
matplotlib==3.10.6
seaborn==0.13.2
joblib==1.5.2
ipykernel==6.30.1
statsmodels==0.14.5
-e .